% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otter.r
\name{execute_script}
\alias{execute_script}
\title{Execute a string as an R script and return the environment from that execution.}
\usage{
execute_script(script, secret)
}
\arguments{
\item{script}{The string to be executed}

\item{secret}{The string to be appended to the name \code{check_results_} as the list name to collect
results}
}
\value{
The global environment after executing the script
}
\description{
Converts a string to an AST and executes that script in a dummy environment for running test cases
against. Transforms all expressions of the form \code{. = otter::check(...)} by replacing the \code{.} with
an index into a list in the environment with name \verb{check_results_\{SECRET\}} to collect the
test_suite_result objects generated from those checks. (This helps to handle variable name collisions
in tests when grading a script.)
}
