% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otter.r
\name{get_case_metadata}
\alias{get_case_metadata}
\title{Get the entry for the test case with name \code{case_name} from \code{test_metadata}}
\usage{
get_case_metadata(test_metadata, case_name)
}
\arguments{
\item{test_metadata}{The parsed test metadata}

\item{case_name}{The name of the desired case}
}
\value{
The configuration for test case CASE_NAME
}
\description{
Get the entry for the test case with name \code{case_name} from \code{test_metadata}
}
