% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otter.r
\name{load_test_metadata}
\alias{load_test_metadata}
\title{Load test suite metadata from a file}
\usage{
load_test_metadata(file)
}
\arguments{
\item{file}{Path to a test suite file}
}
\value{
The parse test suite metadata
}
\description{
Executes the script FILE expression-by-expression and extracts the global variable test_metadata.
This string is run through a YAML parser to construct a list containing the metadata specifications
for that test suite. The global key \code{name} should be defined and \code{cases} should be a list of
dictionaries with keys \code{name} and \code{hidden}. The key \code{case[[int]][["name"]]} should match a name
passed to a call to test_that.
}
\details{
For example, the test suite might have the following contents:\if{html}{\out{<div class="r">}}\preformatted{library(testthat)

test_metadata = "
name: q1
cases:
  - name: q1a
    hidden: false
  - name: q1b
    hidden: true
"

test_that("q1a", \{...\})

test_that("q1b", \{...\})
}\if{html}{\out{</div>}}
}
