% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otter.r
\name{make_secret}
\alias{make_secret}
\title{Randomly generate a string of \code{n_chars} sampled at random from \code{valid_chars}.}
\usage{
make_secret(n_chars, valid_chars)
}
\arguments{
\item{n_chars}{The number of characters in the string; defaults to 6}

\item{valid_chars}{A string of characters to choose from; defaults to all alphanumerals, \code{.}, and
\verb{_}}
}
\value{
The generated string
}
\description{
Randomly generate a string of \code{n_chars} sampled at random from \code{valid_chars}.
}
