% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otter.r
\name{run_gradescope}
\alias{run_gradescope}
\title{Run autograder in a Gradescope container and return the results as a properly-formatted JSON
string}
\usage{
run_gradescope(script_path, secret)
}
\arguments{
\item{script_path}{The path to the script}

\item{secret}{The string to be appended to the name \code{check_results_} as the list name to collect
results (optional)}
}
\value{
The JSON string
}
\description{
Run autograder in a Gradescope container and return the results as a properly-formatted JSON
string
}
